requirements.txt
# theo Research Assistant - Dependencies

# Core Framework
chainlit>=2.0.0

# Environment & Configuration
python-dotenv>=1.0.0

# OpenAI SDK (for OpenRouter API)
openai>=1.0.0

# Async HTTP
httpx>=0.24.0
aiohttp>=3.8.0

# Data handling
pydantic>=2.0.0
